entrega final

+++ Practica 8 Max Mean Dispersion Problem/include/edge.hpp
@@ -30,22 +30,22 @@
 class Edge {
 
 	private:
-		int distance_;
+		float distance_;
 		int vertexA_;
 		int vertexB_;
 
 	public:
 		// Builders & Destroyer
 		Edge ();
-		Edge (int vertexA, int vertexB, int  distance);
+		Edge (int vertexA, int vertexB, float  distance);
 		~Edge ();
 
 		// Getters & Setters
-		int get_Distance (void) const;
+		float get_Distance (void) const;
 		int get_VertexA (void) const;
 		int get_VertexB (void) const;
 
-		void set_Distance (int distance);
+		void set_Distance (float distance);
 		void set_VertexA (int vertexA);
 		void set_VertexB (int vertexB);
 };
